feat: add userLogin e getUser no contexto do usuario
salvando as informações nesse contexto para reutilizar no resto da aplicação
