+++ src/Components/Header.jsx
@@ -2,8 +2,10 @@ import React from 'react';
 import styles from './Header.module.css';
 import { Link } from 'react-router-dom';
 import { ReactComponent as Dogs } from '../Assets/dogs.svg';
+import { UserContext } from '../UserContext';
 
 const Header = () => {
+  const { data } = React.useContext(UserContext);
   return (
     <div>
       <header className={styles.header}>
@@ -11,9 +13,15 @@ const Header = () => {
           <Link className={styles.logo} to="/" aria-label="Dogs - Home">
             <Dogs />
           </Link>
-          <Link className={styles.login} to="/login">
-            Login / Criar
-          </Link>
+          {data ? (
+            <Link className={styles.login} to="/conta">
+              {data.username}
+            </Link>
+          ) : (
+            <Link className={styles.login} to="/login">
+              Login / Criar
+            </Link>
+          )}
         </nav>
       </header>
     </div>
