+++ src/UserContext.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { TOKEN_POST, USER_GET } from './api';
 
 export const UserContext = React.createContext();
 
@@ -8,8 +9,25 @@ export const UserStorage = ({ children }) => {
   const [loading, setLoading] = React.useState(false);
   const [error, setError] = React.useState(null);
 
+  async function getUser(token) {
+    const { url, options } = USER_GET(token);
+    const response = await fetch(url, options);
+    const json = await response.json();
+    setData(json);
+    setLogin(true);
+    console.log(json);
+  }
+
+  async function userLogin(username, password) {
+    const { url, options } = TOKEN_POST({ username, password });
+    const tokenRes = await fetch(url, options);
+    const { token } = await tokenRes.json();
+    window.localStorage.setItem('token', token);
+    getUser(token);
+  }
+
   return (
-    <UserContext.Provider value={{ usuario: 'André' }}>
+    <UserContext.Provider value={{ userLogin, data }}>
       {children}
     </UserContext.Provider>
   );
